requirements.txt
# GUI
PyQt5>=5.15.0

# PostgreSQL
psycopg2-binary>=2.9.0

# Для работы с датами/временными метками
python-dateutil>=2.8.2

# Логирование (если захочешь расширенное логирование)
loguru>=0.7.0

# Опционально: для работы с таблицами в окнах (если будешь использовать Pandas для фильтров/выводов)
pandas>=2.0.0
